refactor: Rename directory
